Add NNAPI's "LegacyHalUtils.h" to Utils.hpp
The NNAPI is in the process of decoupling the NN runtime and utility
code from HIDL dependencies. Utility code that relies on HIDL types is
being relocated to the file "LegacyHalUtils.h".

Currently, `setRunTimePoolInfosFromHidlMemories` is defined in
CpuExecutor.h, but this function is being relocated to LegacyHalUtils.h.
This change adds "LegacyHalUtils.h" to android-nn-driver/Utils.hpp in
preparation for this migration.

Bug: 170289677
Test: mma
Change-Id: If8eb0cf0def8a9103e31c7cf7415544667646006